extract web3auth prompts
